NSKeyedUnarchiver - Rolf Jansen's fix for parsing scientific notation on floats NSConnection subproject - added a lot of empty methods
Introduction of KoreGraphics (KG prefix) classes to the AppKit project. This is primarily a first pass at PDF parsing and rendering. It is not integrated with the rest of the AppKit, it is simply there and compiles.

--HG--
extra : convert_revision : svn%3A43251c19-cd1f-0410-9ec0-95c558c233a8/trunk%407